backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
datasets
pandas
openpyxl
python-docx
PyPDF2
pdfminer.six